feat: Add resource_name variable
We're looking at using this module to deploy multiple sets of runner
types (x86_64 and arm architectures) but within the same conceptual
"environment".

We use the "Environment" tag throughout our tooling, but the constraints
of using the "environment" variable for resource naming mean that we
need to essentially supply different environment names (eg "env-amd64"
and "env-arm64"), even though they are not in different environments.

We also use the dot character (".") in our environment names, which
isn't allowed in some resource names (eg SQS queue name).

This commit adds an additional variable called "resource_name" that if
set, is used to name resources.

This means we can set it alongside the "environment" variable to ensure
we have the correct tags, but resource names are distinct between
multiple uses of this module.

If the "resource_name" variable isn't used, then it's ignored.